add color func.

+++ run.py
@@ -1,4 +1,5 @@
 # -*- coding: utf-8 -*-
+import contextlib
 import os
 import sys
 from datetime import datetime
@@ -6,6 +7,8 @@ from ipaddress import ip_address
 
 from chardet import detect
 from openpyxl import load_workbook
+from openpyxl.styles import PatternFill
+
 import configparser
 
 
@@ -19,24 +22,6 @@ return {*}
 def is_sheet_empty(sheet):
     return sheet.max_row == 1 and sheet.max_column == 1
 
-# deprecated
-def get_row(filename, header_row):
-
-    global name, input_col
-    backnames = ['IP地址', '专网IP','ip地址']
-    for back_name in backnames:
-        try:
-            col = header_row.index(back_name)
-            return col
-        except ValueError:
-            pass
-    # 如果找不到表头，则全部sheet以用户输入的列数为准，且只指定一次
-
-    if input_col == -1:
-        input_col = int(input(f"在'{filename}'中的'{sheet.title}'找不到IP表头,输入IP所在列的列数,或输入0跳过"))
-        if input_col == 0:
-            return -1
-    return input_col-1
 
 '''
 description: 
@@ -47,25 +32,17 @@ return {*}
 def get_ip_row_in_sheet(filename, sheet):
     if is_sheet_empty(sheet):
         return -1
-    global name, input_col
-    backnames = ['IP地址', '专网IP','ip地址']
-    # 在前四行尝试寻找表头
-    max_try_times = 5
+    global backnames
     row_generator = sheet.iter_rows(values_only=True)
-    while(max_try_times > 0):
+    for _ in range(5, 0, -1):
         try:
             header_row = next(row_generator)
         except StopIteration:
             return -1
         for back_name in backnames:
-            try:
-                col = header_row.index(back_name)
-                return col
-            except ValueError:
-                pass
-        max_try_times = max_try_times - 1
-
-    return input_col-1
+            with contextlib.suppress(ValueError):
+                return header_row.index(back_name)
+    return -1
 
 
 '''
@@ -79,8 +56,7 @@ def read_xlsx(files=None):
         try:
             wb = load_workbook(fn, read_only=True)
         except PermissionError:
-            input("文件被占用,无法读取,按任意键退出")
-            sys.exit(1)
+            input("文件被占用,无法读取,关闭后按enter继续")
 
         for sheet in wb.worksheets:
             col_num = get_ip_row_in_sheet(os.path.basename(fn), sheet)
@@ -89,14 +65,10 @@ def read_xlsx(files=None):
             for row in sheet.iter_rows():
                 cell_value = row[col_num].value
                 if cell_value is not None:
-                    try:
+                    with contextlib.suppress(ValueError):
                         ip_address(cell_value)
-                        # if ip_address(cell_value).is_private == False:
                         data.append(cell_value)
-                    except ValueError:
-                        pass
-    data = list(set(data))  # 去重
-    return data
+    return list(set(data))
 
 
 def get_ip_list(xfile, result):
@@ -104,7 +76,6 @@ def get_ip_list(xfile, result):
     with open(result, 'w') as f:
         for item in data:
             f.write(f"{item}\n")
-
 
 
 def count_lines(filename):
@@ -158,28 +129,28 @@ description:
 param {*} file_list
 param {*} bad_ip
 return {*}
+    规则: 在线统计结果放置于表格右侧空白一列
 '''
 def write_result(file_list, ip_file, is_in, not_in):
     for fn in file_list:
         try:
             wb = load_workbook(fn)
         except PermissionError:
-            input("文件被占用,无法读取,按任意键退出")
-            sys.exit(1)
+            input("文件被占用,无法读取,关闭后按enter继续")
 
         for sheet in wb.worksheets:
             col_num = get_ip_row_in_sheet(os.path.basename(fn), sheet)
             if col_num <= -1:
                 continue
             max_col = sheet.max_column
-            new_col = max_col + 1  # 增加 “是否在线” 统计结果列
-            sheet.cell(row=1, column=new_col).value = datetime.now().strftime("是否在线%m%d %H:%M")
+            new_col = max_col + 1  # 增加在线统计结果列
+            sheet.cell(row=1, column=new_col).value = datetime.now().strftime("%m%d %H:%M")
 
             with open(ip_file, "r") as f:
                 ip_list = set(f.read().splitlines())
 
             for row in range(2, sheet.max_row + 1):
-                cell_value = sheet.cell(row=row, column=col_num + 1).value  # sheet.cell 的 column 是从1开始
+                cell_value = sheet.cell(row=row, column=col_num + 1).value  # sheet.cell 的 column 是从1开始,而col_num是从0开始
                 try:
                     ip_address(cell_value)
                 except ValueError:
@@ -193,16 +164,61 @@ def write_result(file_list, ip_file, is_in, not_in):
                 #    sheet.cell(row=row, column=new_col).value = ''
         wb.save(fn)
 
+'''
+description: 
+param {*} write_index_offset 写入结果列与IP地址所在列的index差值
+    定制规则：
+        if cell_value in ip_list 的结果记为I,IP单元格的右3 cell记为R,左3 cell记为L
+        1.R根据I正常填充is_in、not_in
+        2.L填充红色,跳过
+        3.L无填充色/白色且非空,I为真,L.value=is_in
+        4.L无填充色/白色,且空;I为假,L.value=1,填充红色
+'''
+def write_result_color(file_list, ip_file, is_in, not_in, write_index_offset=3):
+    fille = PatternFill("solid",fgColor="FFFF0000")
+    for fn in file_list:
+        try:
+            wb = load_workbook(fn)
+        except PermissionError:
+            input("文件被占用,无法读取,关闭后按enter继续")
+
+        for sheet in wb.worksheets:     # 跳过无IP的sheet
+            col_num = get_ip_row_in_sheet(os.path.basename(fn), sheet)
+            if col_num <= -1:
+                continue
+            
+            with open(ip_file, "r") as f:
+                ip_list = set(f.read().splitlines())
+
+            for row in range(1, sheet.max_row + 1):
+                cell_value = sheet.cell(row=row, column=col_num + 1).value  # sheet.cell 的 column 是从1开始,而col_num是从0开始
+                try:
+                    ip_address(cell_value)
+                except ValueError:
+                    continue
+                
+                bool_is_in = cell_value in ip_list
+                sheet.cell(row=row, column=col_num + 1 + write_index_offset).value  = is_in if bool_is_in else not_in
+
+                fill = sheet.cell(row=row, column=col_num - 2).fill     #涂色位置
+                value = sheet.cell(row=row, column=col_num - 2).value   #涂色格的值
+
+                if fill.fgColor.rgb != 'FFFF0000':          #标准红色跳过
+                    if value is not None:                     #非红,非空
+                        sheet.cell(row=row, column=col_num - 2).value = is_in if bool_is_in else 1
+                    elif not bool_is_in:      #非红,不在list中,空
+                        sheet.cell(row=row, column=col_num - 2).value = 1
+                        sheet.cell(row=row, column=col_num - 2).fill = fille
+                        print(cell_value," 新增在线,加入未上墙")
+        wb.save(fn)
 
 def remove_txt():
     file_path = os.getcwd()
     file_ext = '.txt'
     for path in os.listdir(file_path):
         path_list = os.path.join(file_path, path)
-        if os.path.isfile(path_list):
-            if (os.path.splitext(path_list)[1]) == file_ext:
-                os.remove(path_list)
-
+        if os.path.isfile(path_list) and (os.path.splitext(path_list)[1]) == file_ext:
+            os.remove(path_list)
 
 '''
 description: 对XLSX文件列表,执行IP连通性测试,并回写结果
@@ -210,38 +226,40 @@ param {*} file_list
 return {*}
 '''
 def ip_xlsx_test(file_list, ping_timeout, ping_size, times):
-    get_ip_list(file_list, ip_list_name)
-    # 未测试时bad ip为全部ip
-    with open(ip_list_name, 'r') as file1, open(bad_name, 'w') as file2:
+    get_ip_list(file_list, "ip.txt")
+    bad_name = "bad.txt"
+    with open("ip.txt", 'r') as file1, open(bad_name, 'w') as file2:
         file2.write(file1.read())
-    # ping重复操作
+
     print(f'ping设置:超时{ping_timeout}ms,{ping_size}字节')
     while(times > 0):
-        call_ping(ping_relative_path, bad_name, result_name, ping_timeout, ping_size)
-        get_bad_ip(result_name, bad_name)
-
+        call_ping(r'tools\PingInfoView.exe', bad_name, "result.txt", ping_timeout, ping_size)
+        get_bad_ip("result.txt", bad_name)
         times = times - 1
         print(f'剩余{times}次,',end='')
+
     # 结果回写
-    write_result(file_list, bad_name, "请求超时", "成功完成")
-    input("操作完成,按任意键退出.")
+    # write_result(file_list, bad_name, "超时", "成功")
+    write_result_color(file_list, bad_name, "掉线", "正常")
+
+def main():
+    global config
+    ping_timeout = config.getint('cfg', 'ping_timeout', fallback=5000)    #ping超时时间
+    ping_times = config.getint('cfg', 'ping_times', fallback=25)        #ping尝试次数
+    file_list = [f for f in os.listdir() if f.endswith('.xlsx')]
+    ip_xlsx_test(file_list, ping_timeout, 4, ping_times)
     remove_txt()
 
-# if __name__ == 'main':
-ip_list_name = "ip.txt"
-result_name = "result.txt"
-bad_name = "bad.txt"
-ping_relative_path = r'tools\PingInfoView.exe'
-name = "专网IP"
-input_col = -1  # 自定义IP所在列
-config = configparser.ConfigParser()
-config.read('config/config.ini',encoding="utf-8-sig")
-
-ping_timeout = config.getint('cfg', 'ping_timeout', fallback=5000)    #ping超时时间
-ping_times = config.getint('cfg', 'ping_times', fallback=25)        #ping尝试次数
-ping_size = 4
-
-print(os.getcwd())
-file_list = [f for f in os.listdir() if f.endswith('.xlsx')]
-ip_xlsx_test(file_list, ping_timeout, ping_size, ping_times)
-sys.exit(0)
+if __name__ == '__main__':
+    print(os.getcwd())
+    config = configparser.ConfigParser()
+    config.read('config/config.ini',encoding="utf-8-sig")
+    head_name = config.get('cfg', 'head_name', fallback="IP")        #定义表头
+    backnames = ['专网IP','IP地址']
+    backnames.insert(0,head_name)
+    try:
+        main()
+        input("操作完成,按enter退出")
+        sys.exit(0)
+    except Exception as e:
+        print(e)
